+++ ensembl_dbcopy/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'ensembl_dbcopy.apps.EnsemblDbcopyConfig'
