+++ ensembl_dbcopy/apps.py
@@ -17,3 +17,6 @@ from django.apps import AppConfig
 class EnsemblDbcopyConfig(AppConfig):
     name = 'ensembl_dbcopy'
     verbose_name = "Ensembl DB Copy"
+
+    def ready(self):
+        import ensembl_dbcopy.signals
